[Done]- Secured Supplements

+++ src/app/api/supplement-items/route.js
@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
 import { SupplementItem } from "../../models/supplement-item";
+import {isAdmin} from '../auth/[...nextauth]/route'
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
@@ -12,9 +13,12 @@ export async function POST(req) {
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
   const { _id, ...data } = await req.json();
-
-  await SupplementItem.findByIdAndUpdate(_id, data);
-  return Response.json(true);
+  if (await isAdmin()) {
+    await SupplementItem.findByIdAndUpdate(_id, data);
+    return Response.json(true);
+  } else {
+    return Response.json([]);
+  }
 }
 
 export async function GET() {
@@ -26,6 +30,9 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get("_id");
+  if(await isAdmin ()) {
   await SupplementItem.deleteOne({ _id });
+  }
+
   return Response.json(true);
 }
